+++ simple-game-service/src/main/java/com/samuilolegovich/repository/ConditionRepo.java
@@ -4,9 +4,11 @@ import com.samuilolegovich.domain.Condition;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 
 @Repository
 public interface ConditionRepo extends CrudRepository<Condition, Long> {
-    Condition findByBet(long bet);
+    Condition findByBet(BigDecimal bet);
     Condition findById(int id);
 }
